+++ src/adt/IAdtDataElement.ts
@@ -17,12 +17,7 @@ export interface ICreateDataElementParams {
   masterSystem?: string;
   responsible?: string;
   masterLanguage?: string;
-  type_kind?:
-    | 'domain'
-    | 'predefinedAbapType'
-    | 'refToPredefinedAbapType'
-    | 'refToDictionaryType'
-    | 'refToClifType';
+  type_kind?: DataElementTypeKind;
   type_name?: string;
   data_type?: string;
   length?: number;
@@ -46,12 +41,7 @@ export interface IUpdateDataElementParams {
   description?: string;
   package_name: string;
   transport_request?: string;
-  type_kind?:
-    | 'domain'
-    | 'predefinedAbapType'
-    | 'refToPredefinedAbapType'
-    | 'refToDictionaryType'
-    | 'refToClifType';
+  type_kind?: DataElementTypeKind;
   type_name?: string;
   data_type?: string;
   length?: number;